refactor(ios): share Agent Pro detail components
